test : 문자열 관련 테스트 분리
