+++ src/test/java/stringCalculator/StringCalculatorTest.java
@@ -22,14 +22,6 @@ public class StringCalculatorTest {
     }
 
     @Test
-    @DisplayName("문자열 구분자")
-    void separatorTest(){
-        String inputString = "1,2;3;";
-        String[] stringArray = StringUtil.complicatedSeparator(inputString);
-        assertThat(stringArray).contains("1","2","3");
-    }
-
-    @Test
     @DisplayName("덧셈")
     void plusCalculateTest(){
         String[] numbers = {"1","2","3"};
@@ -53,21 +45,5 @@ public class StringCalculatorTest {
                 .isInstanceOf(NumberFormatException.class);
     }
 
-    @Test
-    @DisplayName("커스텀 구분자")
-    void customSeparatorTest(){
-        String customSeparatorContainString = "//a\n1a2a3";
-        String[] strings = StringUtil.complicatedSeparator(customSeparatorContainString);
-        assertThat(strings).contains("1","2","3");
-    }
-
-
-    @Test
-    @DisplayName("커스텀 구분자와 쉼표, 콜론 혼합")
-    void complicatedStringTest(){
-        String complicatedString = "//a\n1a2,3;4";
-        String[] strings = StringUtil.complicatedSeparator(complicatedString);
-        assertThat(strings).contains("1","2","3","4");
-    }
 
 }
